Move server code to server directory
